refact model runner
